Split typed feedback tests out of runtime module
